+++ modules/terraform-zscc-ccvm-aws/variables.tf
@@ -107,3 +107,9 @@ variable "ami_id" {
   type        = list(string)
   description = "AMI ID(s) to be used for deploying Cloud Connector appliances. Ideally all VMs should be on the same AMI ID as templates always pull the latest from AWS Marketplace. This variable is provided if a customer desires to override/retain an old ami for existing deployments rather than upgrading and forcing a replacement. It is also inputted as a list to facilitate if a customer desired to manually upgrade select CCs deployed based on the cc_count index"
 }
+
+variable "imdsv2_enabled" {
+  type        = bool
+  description = "true/false whether to force IMDSv2 only for instance bring up. Default is true"
+  default     = true
+}